+++ encon-handler/src/main/java/io/appulse/encon/handler/message/matcher/Matchers.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors..
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.appulse.encon.handler.mock;
 
-import static io.appulse.encon.handler.mock.ArgumentMatcher.ANY;
-import static io.appulse.encon.handler.mock.ArgumentMatcher.NOT_NULL;
-import static io.appulse.encon.handler.mock.ArgumentMatcher.NULL;
-import static io.appulse.encon.handler.mock.ThreadLocalStorage.stack;
-import static io.appulse.encon.handler.mock.Primitives.defaultValue;
+package io.appulse.encon.handler.message.matcher;
+
+import static io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.ANY;
+import static io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.NOT_NULL;
+import static io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.NULL;
+import static io.appulse.encon.handler.message.matcher.Primitives.defaultValue;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
@@ -30,17 +30,17 @@ import java.util.Map;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-import io.appulse.encon.handler.mock.ArgumentMatcher.Contains;
-import io.appulse.encon.handler.mock.ArgumentMatcher.EndsWith;
-import io.appulse.encon.handler.mock.ArgumentMatcher.Equals;
-import io.appulse.encon.handler.mock.ArgumentMatcher.InstanceOf;
-import io.appulse.encon.handler.mock.ArgumentMatcher.Matches;
-import io.appulse.encon.handler.mock.ArgumentMatcher.ReflectionEquals;
-import io.appulse.encon.handler.mock.ArgumentMatcher.Same;
-import io.appulse.encon.handler.mock.ArgumentMatcher.StartsWith;
+import io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.Contains;
+import io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.EndsWith;
+import io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.Equals;
+import io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.InstanceOf;
+import io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.Matches;
+import io.appulse.encon.handler.message.matcher.MethodArgumentMatcher.StartsWith;
 
 /**
+ * The set of pattern matching helpers.
  *
+ * @since 1.4.0
  * @author alabazin
  */
 public final class Matchers {
@@ -62,14 +62,14 @@ public final class Matchers {
   /**
    * Matches any object of given type, excluding nulls.
    *
-   * @param <T> the accepted type
+   * @param <T>  the accepted type
    *
    * @param type the class of the accepted type
    *
    * @return {@code null}
    */
   public static <T> T any (Class<T> type) {
-    reportMatcher(new InstanceOf.VarArgAware(type));
+    reportMatcher(new InstanceOf(type));
     return defaultValue(type);
   }
 
@@ -94,7 +94,7 @@ public final class Matchers {
    */
   public static boolean anyBoolean () {
     reportMatcher(new InstanceOf(Boolean.class));
-    return false;
+    return defaultValue(Boolean.class);
   }
 
   /**
@@ -104,7 +104,7 @@ public final class Matchers {
    */
   public static byte anyByte () {
     reportMatcher(new InstanceOf(Byte.class));
-    return 0;
+    return defaultValue(Byte.class);
   }
 
   /**
@@ -114,7 +114,7 @@ public final class Matchers {
    */
   public static char anyChar () {
     reportMatcher(new InstanceOf(Character.class));
-    return 0;
+    return defaultValue(Character.class);
   }
 
   /**
@@ -124,7 +124,7 @@ public final class Matchers {
    */
   public static short anyShort () {
     reportMatcher(new InstanceOf(Short.class));
-    return 0;
+    return defaultValue(Short.class);
   }
 
   /**
@@ -134,7 +134,7 @@ public final class Matchers {
    */
   public static int anyInt () {
     reportMatcher(new InstanceOf(Integer.class));
-    return 0;
+    return defaultValue(Integer.class);
   }
 
   /**
@@ -144,7 +144,7 @@ public final class Matchers {
    */
   public static long anyLong () {
     reportMatcher(new InstanceOf(Long.class));
-    return 0;
+    return defaultValue(Long.class);
   }
 
   /**
@@ -154,7 +154,7 @@ public final class Matchers {
    */
   public static float anyFloat () {
     reportMatcher(new InstanceOf(Float.class));
-    return 0;
+    return defaultValue(Float.class);
   }
 
   /**
@@ -164,7 +164,7 @@ public final class Matchers {
    */
   public static double anyDouble () {
     reportMatcher(new InstanceOf(Double.class));
-    return 0;
+    return defaultValue(Double.class);
   }
 
   /**
@@ -174,7 +174,7 @@ public final class Matchers {
    */
   public static String anyString () {
     reportMatcher(new InstanceOf(String.class));
-    return "";
+    return defaultValue(String.class);
   }
 
   /**
@@ -248,7 +248,7 @@ public final class Matchers {
    */
   public static boolean eq (boolean value) {
     reportMatcher(new Equals(value));
-    return false;
+    return defaultValue(Boolean.class);
   }
 
   /**
@@ -260,7 +260,7 @@ public final class Matchers {
    */
   public static byte eq (byte value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Byte.class);
   }
 
   /**
@@ -272,7 +272,7 @@ public final class Matchers {
    */
   public static char eq (char value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Character.class);
   }
 
   /**
@@ -284,7 +284,7 @@ public final class Matchers {
    */
   public static short eq (short value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Short.class);
   }
 
   /**
@@ -296,7 +296,7 @@ public final class Matchers {
    */
   public static int eq (int value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Integer.class);
   }
 
   /**
@@ -308,7 +308,7 @@ public final class Matchers {
    */
   public static long eq (long value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Long.class);
   }
 
   /**
@@ -320,7 +320,7 @@ public final class Matchers {
    */
   public static float eq (float value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Float.class);
   }
 
   /**
@@ -332,7 +332,7 @@ public final class Matchers {
    */
   public static double eq (double value) {
     reportMatcher(new Equals(value));
-    return 0;
+    return defaultValue(Double.class);
   }
 
   /**
@@ -346,42 +346,7 @@ public final class Matchers {
    */
   public static <T> T eq (T value) {
     reportMatcher(new Equals(value));
-    return value == null
-           ? null
-           : (T) defaultValue(value.getClass());
-  }
-
-  /**
-   * Object argument that is reflection-equal to the given value with support for excluding
-   * selected fields from a class.
-   *
-   * @param value         the given value
-   *
-   * @param excludeFields fields to exclude, if field does not exist it is ignored
-   *
-   * @param <T>           the accepted type
-   *
-   * @return {@code null}
-   */
-  public static <T> T refEq (T value, String... excludeFields) {
-    reportMatcher(new ReflectionEquals(value, excludeFields));
-    return null;
-  }
-
-  /**
-   * Object argument that is the same as the given value.
-   *
-   * @param <T>   the type of the object, it is passed through to prevent casts
-   *
-   * @param value the given value
-   *
-   * @return {@code null}
-   */
-  public static <T> T same (T value) {
-    reportMatcher(new Same(value));
-    return value == null
-           ? null
-           : (T) defaultValue(value.getClass());
+    return (T) defaultValue(value.getClass());
   }
 
   /**
@@ -417,7 +382,7 @@ public final class Matchers {
    */
   public static String contains (String substring) {
     reportMatcher(new Contains(substring));
-    return "";
+    return defaultValue(String.class);
   }
 
   /**
@@ -429,7 +394,7 @@ public final class Matchers {
    */
   public static String regex (String pattern) {
     reportMatcher(new Matches(pattern));
-    return "";
+    return defaultValue(String.class);
   }
 
   /**
@@ -441,7 +406,7 @@ public final class Matchers {
    */
   public static String regex (Pattern pattern) {
     reportMatcher(new Matches(pattern));
-    return "";
+    return defaultValue(String.class);
   }
 
   /**
@@ -453,7 +418,7 @@ public final class Matchers {
    */
   public static String endsWith (String suffix) {
     reportMatcher(new EndsWith(suffix));
-    return "";
+    return defaultValue(String.class);
   }
 
   /**
@@ -465,7 +430,7 @@ public final class Matchers {
    */
   public static String startsWith (String prefix) {
     reportMatcher(new StartsWith(prefix));
-    return "";
+    return defaultValue(String.class);
   }
 
   /**
@@ -477,7 +442,7 @@ public final class Matchers {
    *
    * @return {@code null}
    */
-  public static <T> T argThat (ArgumentMatcher<T> matcher) {
+  public static <T> T argThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
     return null;
   }
@@ -489,9 +454,9 @@ public final class Matchers {
    *
    * @return {@code false}
    */
-  public static boolean booleanThat (ArgumentMatcher<Boolean> matcher) {
+  public static boolean booleanThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return false;
+    return defaultValue(Boolean.class);
   }
 
   /**
@@ -501,9 +466,9 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static byte byteThat (ArgumentMatcher<Byte> matcher) {
+  public static byte byteThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Byte.class);
   }
 
   /**
@@ -513,9 +478,9 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static char charThat (ArgumentMatcher<Character> matcher) {
+  public static char charThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Character.class);
   }
 
   /**
@@ -525,9 +490,9 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static short shortThat (ArgumentMatcher<Short> matcher) {
+  public static short shortThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Short.class);
   }
 
   /**
@@ -537,9 +502,9 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static int intThat (ArgumentMatcher<Integer> matcher) {
+  public static int intThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Integer.class);
   }
 
   /**
@@ -549,9 +514,9 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static long longThat (ArgumentMatcher<Long> matcher) {
+  public static long longThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Long.class);
   }
 
   /**
@@ -561,9 +526,9 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static float floatThat (ArgumentMatcher<Float> matcher) {
+  public static float floatThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Float.class);
   }
 
   /**
@@ -573,16 +538,15 @@ public final class Matchers {
    *
    * @return {@code 0}
    */
-  public static double doubleThat (ArgumentMatcher<Double> matcher) {
+  public static double doubleThat (MethodArgumentMatcher matcher) {
     reportMatcher(matcher);
-    return 0;
+    return defaultValue(Double.class);
   }
 
-  private static void reportMatcher (ArgumentMatcher<?> argumentMatcher) {
-    stack().push(argumentMatcher);
+  private static void reportMatcher (MethodArgumentMatcher argumentMatcher) {
+    ThreadLocalStorage.add(argumentMatcher);
   }
 
-  // https://github.com/mockito/mockito/blob/release/2.x/src/main/java/org/mockito/ArgumentMatchers.java#L1103
   private Matchers () {
   }
 }